pyproject.toml
[project]
name = "upnext"
version = "0.0.1"
description = "A simple program to keep track of which episode you're on in series on the file system."
authors = [
  {name = "Peter Kementzey", email = "your@email.com"},
]
readme = "README.md"

requires-python = ">= 3.8"

dependencies = [
    "typer==0.11.0",
    "ruamel.yaml==0.18.6"
]
